UPDATE: Refactoring PrimesGenerator
Rename as PrimesTable and use it as a service
