+++ Classlib/Interfaces/Services/IPrimesTable.php
@@ -1,0 +1,12 @@
+<?php
+namespace DevSpace\Interfaces\Services;
+
+
+interface IPrimesTable
+{
+    /**
+     * @param int $size
+     * @return array
+     */
+    public function getPrimesTable($size);
+}
